Added new feature which allows for time-varying production or degradation rates for specified genes
The user provides a dataframe whose column names are the parameters that they want to adjust during the simulation. The columns are the fixed values that the simulation interpolates through. The number of values provided can be any number, as long as the same amount is provided per parameter. The first value in each vector of the dataframe is treated as t=0 and the last value is treated as t=simulationTime. The simulation currently uses step function interpolation between the values and multiplies the specified parameters by the interpolated values at each point in time.